requirements.txt
Package            Version
------------------ -----------
aiofiles           23.2.1
aiohappyeyeballs   2.6.1
aiohttp            3.11.16
aiosignal          1.3.2
annotated-types    0.7.0
anyio              4.9.0
attrs              25.3.0
audioop-lts        0.2.1
azure-core         1.32.0
azure-cosmos       4.9.0
certifi            2025.1.31
charset-normalizer 3.4.1
click              8.1.8
colorama           0.4.6
distro             1.9.0
fastapi            0.115.12
ffmpy              0.5.0
filelock           3.18.0
frozenlist         1.5.0
fsspec             2025.3.2
gradio             5.23.3
gradio_client      1.8.0
groovy             0.1.2
h11                0.14.0
httpcore           1.0.7
httpx              0.28.1
huggingface-hub    0.30.1
idna               3.10
ijson              3.3.0
Jinja2             3.1.6
jiter              0.9.0
markdown-it-py     3.0.0
MarkupSafe         3.0.2
mdurl              0.1.2
multidict          6.3.2
nest-asyncio       1.6.0
numpy              2.2.4
openai             1.70.0
orjson             3.10.16
packaging          24.2
pandas             2.2.3
pillow             11.1.0
pip                24.3.1
propcache          0.3.1
pydantic           2.11.2
pydantic_core      2.33.1
pydub              0.25.1
Pygments           2.19.1
python-dateutil    2.9.0.post0
python-dotenv      1.1.0
python-multipart   0.0.20
pytz               2025.2
PyYAML             6.0.2
requests           2.32.3
rich               14.0.0
ruff               0.11.3
safehttpx          0.1.6
semantic-version   2.10.0
shellingham        1.5.4
six                1.17.0
sniffio            1.3.1
starlette          0.46.1
tenacity           9.1.2
tomlkit            0.13.2
tqdm               4.67.1
typer              0.15.2
typing_extensions  4.13.1
typing-inspection  0.4.0
tzdata             2025.2
urllib3            2.3.0
uvicorn            0.34.0
websockets         15.0.1
yarl               1.18.3
